# Optimize some string.Substring usages