--- a/src/Servers/HttpSys/src/UrlPrefix.cs
+++ b/src/Servers/HttpSys/src/UrlPrefix.cs
@@ -127,11 +127,11 @@ public class UrlPrefix
         }
 
         scheme = whole.Substring(0, schemeDelimiterEnd);
-        var portString = whole.Substring(hostDelimiterEnd, pathDelimiterStart - hostDelimiterEnd); // The leading ":" is included
+        var portString = whole.AsSpan(hostDelimiterEnd, pathDelimiterStart - hostDelimiterEnd); // The leading ":" is included
         int portValue;
-        if (!string.IsNullOrEmpty(portString))
+        if (!portString.IsEmpty)
         {
-            var portValueString = portString.Substring(1); // Trim the leading ":"
+            var portValueString = portString.Slice(1); // Trim the leading ":"
             if (int.TryParse(portValueString, NumberStyles.Integer, CultureInfo.InvariantCulture, out portValue))
             {
                 host = whole.Substring(hostDelimiterStart, hostDelimiterEnd - hostDelimiterStart);