@danmoseley 2022-10-03
2 and 3 are clearly beneficial. 

For 1 as you say it is not a win if there is nothing to trim, and probably a win if there is (it's at least less allocating)

@stephentoub  did we consider an analyzer for 3? also, does it surprise you that for 1, the generated code is [so much larger?](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgAYBYAKDIEYabiBmAAjNYA1WBvR2tsTpJ2dcqwCCdABRDxASwB2ABwCuGAJQ0+1VnvYB2VkrUYAdBNwBlZdkXSNZgCpR5+B84hWMrxQHMHAG4aAF9+JkFhUXEJUlkxYxV1LWodfUNE02dXdw1g6jDdfVCgA===)