@stephentoub 2022-10-04
> did we consider an analyzer for 3?

What's an example of 3?

> does it surprise you that for 1, the generated code is [so much larger?](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgAYBYAKDIEYabiBmAAjNYA1WBvR2tsTpJ2dcqwCCdABRDxASwB2ABwCuGAJQ0+1VnvYB2VkrUYAdBNwBlZdkXSNZgCpR5+B84hWMrxQHMHAG4aAF9+JkFhUXEJUlkxYxV1LWodfUNE02dXdw1g6jDdfVCgA===)

Not really; the method is making three method calls instead of one, and one of those (the `AsSpan()`) will get inlined.