@meziantou 2022-10-04
@danmoseley I revert all changes for the first identified case. I expect most usages to be already trimmed in the common case.